#55: Batch results for IPC - separate perf file
